Moved timer to internal package.
