--- epoxy-adapter/src/main/java/com/airbnb/epoxy/NoOpTimer.java
@@ -1,13 +1,0 @@
-package com.airbnb.epoxy;
-
-class NoOpTimer implements Timer {
-  @Override
-  public void start() {
-
-  }
-
-  @Override
-  public void stop(String message) {
-
-  }
-}
