--- epoxy-adapter/src/main/java/com/airbnb/epoxy/Timer.java
@@ -1,6 +1,0 @@
-package com.airbnb.epoxy;
-
-interface Timer {
-  void start();
-  void stop(String message);
-}
