+++ epoxy-adapter/src/main/java/com/airbnb/epoxy/internal/Timer.java
@@ -1,0 +1,10 @@
+package com.airbnb.epoxy.internal;
+
+import android.support.annotation.RestrictTo;
+import android.support.annotation.RestrictTo.Scope;
+
+@RestrictTo(Scope.LIBRARY_GROUP)
+public interface Timer {
+  void start();
+  void stop(String message);
+}
